refactor(state): extract tar listing helper
